requirements.txt
numpy==1.19.0
Pillow==7.2.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.48.2
matplotlib==3.3.0
streamlit==0.64.0